requirements.txt
Django
pytz
djangorestframework
django-rest-passwordreset
flake8
psycopg2-binary
pyyaml
ujson
django-debug-toolbar
requests
redis
celery
drf_spectacular